Modularize Shares folder into views, components, services
