+++ frontend/src/modules/Shares/components/RevokeShareItemsModal.js
@@ -6,13 +6,13 @@ import { useSnackbar } from 'notistack';
 import PropTypes from 'prop-types';
 import React, { useCallback, useEffect, useState } from 'react';
 import { useParams } from 'react-router-dom';
-import { Defaults } from '../../../../design';
-import { SET_ERROR, useDispatch } from '../../../../globalErrors';
+import { Defaults } from 'design';
+import { SET_ERROR, useDispatch } from 'globalErrors';
 import {
   getShareObject,
   revokeItemsShareObject,
   useClient
-} from '../../../../services';
+} from 'services';
 
 const RevokeShareItemsModal = (props) => {
   const client = useClient();
