+++ frontend/src/modules/Shares/components/index.js
@@ -1,0 +1,4 @@
+export * from './AddShareItemModal';
+export * from './RevokeShareItemsModal';
+export * from './ShareOutboxList';
+export * from './ShareOutboxListItem';
