+++ frontend/src/modules/Shares/services/approveShareObject.js
