+++ frontend/src/modules/Shares/services/deleteShareObject.js
