+++ frontend/src/modules/Shares/services/getShareObject.js
