+++ frontend/src/modules/Shares/services/index.js
@@ -1,0 +1,10 @@
+export * from './addSharedItem';
+export * from './approveShareObject';
+export * from './deleteShareObject';
+export * from './getShareObject';
+export * from './getShareRequestsFromMe';
+export * from './getShareRequestsToMe';
+export * from './rejectShareObject';
+export * from './removeSharedItem';
+export * from './revokeItemsShareObject';
+export * from './submitApproval';
