+++ frontend/src/modules/Shares/services/rejectShareObject.js
