+++ frontend/src/modules/Shares/views/ShareList.js
