+++ frontend/src/modules/Shares/views/ShareView.js
