+++ frontend/src/services/graphql/ShareObject/index.js
@@ -1,11 +1,1 @@
-export * from './addSharedItem';
-export * from './approveShareObject';
 export * from './createShareObject';
-export * from './deleteShareObject';
-export * from './getShareObject';
-export * from './getShareRequestsFromMe';
-export * from './getShareRequestsToMe';
-export * from './rejectShareObject';
-export * from './removeSharedItem';
-export * from './revokeItemsShareObject';
-export * from './submitApproval';
